refactor(model): Lotto 검증 추가
